requirements.txt
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
Django==1.11.8
django-oauth-toolkit==1.0.0
djangorestframework==3.7.7
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
oauthlib==2.0.7
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2018.3
redis==2.10.6
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
